requirements.txt
minio
grpcio
grpcio-health-checking